+++ backend/tests/unit/danswer/connectors/cross_connector_utils/test_table.html
@@ -1,0 +1,39 @@
+<p>This page is to ensure we’re able to parse a table into a tsv</p>
+<table data-table-width="760" data-layout="default" ac:local-id="3ad64d9f-01f1-4f78-876e-0fdf84e826a6">
+   <tbody>
+      <tr>
+         <th>
+            <p><strong>hello</strong></p>
+         </th>
+         <th>
+            <p><strong>there</strong></p>
+         </th>
+         <th>
+            <p><strong>general</strong></p>
+         </th>
+      </tr>
+      <tr>
+         <td>
+            <p>kenobi</p>
+         </td>
+         <td>
+            <p>a</p>
+         </td>
+         <td>
+            <p>b</p>
+         </td>
+      </tr>
+      <tr>
+         <td>
+            <p>c</p>
+         </td>
+         <td>
+            <p>d</p>
+         </td>
+         <td>
+            <p>e</p>
+         </td>
+      </tr>
+   </tbody>
+</table>
+<p />
